pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "E-commerce microservices backend with FastAPI and MongoDB"
readme = "README.md"
authors = [
    { name = "DevHusnainAi", email = "syedhussnaintirmizi@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "beanie>=1.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
